abstract routes
