+++ packages/read-and-create-calendar-events/server/node-express/server.js
@@ -3,12 +3,12 @@ const cors = require('cors');
 const dotenv = require('dotenv');
 const { mockDb } = require('./utils/mock-db');
 const { prettyPrintJSON } = require('./utils/formatting');
+const route = require('./route');
 
 const Nylas = require('nylas');
 const { WebhookTriggers } = require('nylas/lib/models/webhook');
 const { Scope } = require('nylas/lib/models/connect');
 const { ServerBindings } = require('nylas/lib/config');
-const { default: Event } = require('nylas/lib/models/event');
 
 dotenv.config();
 
@@ -69,100 +69,51 @@ const startExpress = () => {
     clientUri,
   });
 
-  // Handle when an account gets connected
-  expressBinding.on(WebhookTriggers.AccountConnected, (payload) => {
-    console.log(
-      'Webhook trigger received, account connected. Details: ',
-      prettyPrintJSON(payload.objectData)
-    );
-  });
+  if (process.env.NODE_ENV === 'development') {
+    // Handle when an account gets connected
+    expressBinding.on(WebhookTriggers.AccountConnected, (payload) => {
+      console.log(
+        'Webhook trigger received, account connected. Details: ',
+        prettyPrintJSON(payload.objectData)
+      );
+    });
 
-  // Handle when a calendar event is created
-  expressBinding.on(WebhookTriggers.EventCreated, (payload) => {
-    console.log(
-      'Webhook trigger received, calendar event created. Details: ',
-      prettyPrintJSON(payload.objectData)
-    );
-  });
+    // Handle when a calendar event is created
+    expressBinding.on(WebhookTriggers.EventCreated, (payload) => {
+      console.log(
+        'Webhook trigger received, calendar event created. Details: ',
+        prettyPrintJSON(payload.objectData)
+      );
+    });
 
-  // Mount the express middleware to your express app
-  const nylasMiddleware = expressBinding.buildMiddleware();
-  app.use('/nylas', nylasMiddleware);
+    // Mount the express middleware to your express app
+    const nylasMiddleware = expressBinding.buildMiddleware();
+    app.use('/nylas', nylasMiddleware);
 
-  // Add route for getting 20 latest calendar events
-  app.get('/nylas/read-events', async (req, res) => {
-    if (!req.headers.authorization) {
-      return res.json('Unauthorized');
-    }
-
-    const user = await mockDb.findUser(req.headers.authorization);
-    if (!user) {
-      return res.sendStatus(401);
-    }
-
-    const events = await nylasClient
-      .with(user.accessToken)
-      .events.list({
-        starts_after: Math.floor(new Date().getTime() / 1000),
-        limit: 20,
-        ...(req.query.calendarId && { calendar_id: req.query.calendarId }),
-      })
-      .then((events) => events);
-
-    return res.json(events);
-  });
+    // Start the Nylas webhook
+    expressBinding
+      .startDevelopmentWebsocket()
+      .then((webhookDetails) =>
+        console.log(
+          'Webhook tunnel registered. Webhook ID: ' + webhookDetails.id
+        )
+      );
+  }
 
   // Add route for getting 20 latest calendar events
-  app.get('/nylas/read-calendars', async (req, res) => {
-    if (!req.headers.authorization) {
-      return res.json('Unauthorized');
-    }
+  app.get('/nylas/read-events', (...args) =>
+    route.readEvents(...args, nylasClient)
+  );
 
-    const user = await mockDb.findUser(req.headers.authorization);
-    if (!user) {
-      return res.json('Unauthorized');
-    }
-
-    const calendars = await nylasClient
-      .with(user.accessToken)
-      .calendars.list()
-      .then((calendars) => calendars);
-
-    return res.json(calendars);
-  });
+  // Add route for getting 20 latest calendar events
+  app.get('/nylas/read-calendars', (...args) =>
+    route.readCalendars(...args, nylasClient)
+  );
 
   // Add route for creating calendar events
-  app.post('/nylas/create-events', async (req, res) => {
-    if (!req.headers.authorization) {
-      return res.json('Unauthorized');
-    }
-
-    const user = await mockDb.findUser(req.headers.authorization);
-    if (!user) {
-      return res.json('Unauthorized');
-    }
-
-    const { calendarId, title, description, startTime, endTime } = req.body;
-    if (!calendarId || !title || !startTime || !endTime) {
-      return res.json(
-        'Missing required fields: calendarId, title, starTime or endTime'
-      );
-    }
-
-    const nylas = nylasClient.with(user.accessToken);
-
-    const event = new Event(nylas);
-
-    event.calendarId = calendarId;
-    event.title = title;
-    event.description = description;
-    event.when.startTime = startTime;
-    event.when.endTime = endTime;
-
-    event.save();
-
-    return res.json(event);
-  });
+  app.post('/nylas/create-events', (...args) =>
+    route.createEvents(...args, nylasClient)
+  );
 
   // Start listening on port 9000
   app.listen(port, () => console.log('App listening on port ' + port));
